# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "BookNonUse": {
        "properties": {
          "from": {
            "format": "date",
            "type": "string"
          },
          "to": {
            "format": "date",
            "type": "string"
          }
        },
        "required": [
          "from",
          "to"
        ],
        "type": "object"
      },
      "BookParty": {
        "properties": {
          "name": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "share": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "role",
          "share"
        ],
        "type": "object"
      },
      "BookUnitEntry": {
        "properties": {
          "builtAreaM2": {
            "type": "string"
          },
          "complete": {
            "type": "boolean"
          },
          "designation": {
            "type": "string"
          },
          "householdCount": {
            "format": "int32",
            "type": "integer"
          },
          "idealParts": {
            "type": "string"
          },
          "nonUse": {
            "items": {
              "$ref": "#/components/schemas/BookNonUse"
            },
            "type": "array"
          },
          "parties": {
            "items": {
              "$ref": "#/components/schemas/BookParty"
            },
            "type": "array"
          },
          "unitId": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "complete",
          "designation",
          "householdCount",
          "idealParts",
          "nonUse",
          "parties",
          "unitId"
        ],
        "type": "object"
      },
...
      "CondominiumBook": {
        "properties": {
          "asOf": {
            "format": "date",
            "type": "string"
          },
          "complete": {
            "type": "boolean"
          },
          "entranceId": {
            "format": "uuid",
            "type": "string"
          },
          "units": {
            "items": {
              "$ref": "#/components/schemas/BookUnitEntry"
            },
            "type": "array"
          }
        },
        "required": [
          "asOf",
          "complete",
          "entranceId",
          "units"
        ],
        "type": "object"
      },
...
      "RetentionApplied": {
        "properties": {
          "animalPassportsCleared": {
            "format": "int32",
            "type": "integer"
          },
          "householdUnlinked": {
            "format": "int32",
            "type": "integer"
          },
          "on": {
            "format": "date",
            "type": "string"
          }
        },
        "required": [
          "animalPassportsCleared",
          "householdUnlinked",
          "on"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Generated from the running code (ADR-013): it states what runs, nothing more. Each operation names the rules it serves in `x-rules`. Operations designed but not built yet are listed in `tools/build_openapi.py` (PLANNED), not here.",
    "summary": "Управление на етажна собственост по ЗУЕС",
    "title": "domuvai api",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/registry/entrances/{entranceId}/book": {
      "get": {
        "operationId": "get_registry_entrances_entranceId_book",
        "parameters": [
          {
            "in": "path",
            "name": "entranceId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "on",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CondominiumBook"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object"
                }
              }
            },
            "description": "Not Found"
          }
        },
        "summary": "Read the Book of the Condominium as of a date",
        "tags": [
          "registry"
        ],
        "x-rules": [
          "PM-BOOK-001",
          "PM-BOOK-002"
        ]
      }
    },
...
    "/api/registry/entrances/{entranceId}/book/retention": {
      "post": {
        "operationId": "post_registry_entrances_entranceId_book_retention",
        "parameters": [
          {
            "in": "path",
            "name": "entranceId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetentionApplied"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object"
                }
              }
            },
            "description": "Not Found"
          }
        },
        "summary": "Anonymise residents' links to named persons and animals' passport numbers past their retention window — today in Sofia, irreversible",
        "tags": [
          "registry"
        ],
        "x-rules": [
          "PM-BOOK-010"
        ]
      }
    },
...
  },
...
}